Extract GameRescanCatalogUpdater from GameRescanService
Peel trophy catalog synchronization (title, group, and trophy refresh
from PSN) out of the admin rescan orchestrator into a focused class.

GameRescanService now delegates catalog updates and keeps worker auth,
stat recalculation, set-version updates, and audit recording. Adds
GameRescanGroupDataFetcher for testable group-data injection and SQLite
tests for version gating, title diffs, and PSVR2 platform preservation.
